refactor: extract loraDatasetSubject leaf to break caption/generate import cycle (#5917)
